robot-gateway/requirements.txt
pytest==7.2.1
pySerialTransfer==2.1.7
paho-mqtt==1.6.1
protobuf==5.27.3
case-converter==1.1.0
